moves createTaskScript to its own file
